+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * [#1159](https://github.com/bbatsov/rubocop/issues/1159): More rigid conditions for when `attr` is considered an offense. ([@jonas054][])
 * [#1167](https://github.com/bbatsov/rubocop/issues/1167): Fix handling of parameters spanning multiple lines in `TrailingComma`. ([@jonas054][])
 * [#1169](https://github.com/bbatsov/rubocop/issues/1169): Fix handling of ternary op conditions in `ParenthesesAroundCondition`. ([@bbatsov][])
+* [#1147](https://github.com/bbatsov/rubocop/issues/1147): WordArray checks arrays with special characters. ([@camilleldn][])
 
 ## 0.23.0 (02/06/2014)
 
@@ -995,3 +996,4 @@
 [@wndhydrnt]: https://github.com/wndhydrnt
 [@ggilder]: https://github.com/ggilder
 [@salbertson]: https://github.com/salbertson
+[@camilleldn]: https://github.com/camilleldn